1. Keyword seq 키 추가 2. Reader class  generic 제거
Former-commit-id: 9e608f8602cbdbc882aa4d988d501855b0d1807a

+++ src/test/java/daon/analysis/ko/dict/test/Dictionary.java
@@ -31,7 +31,7 @@ public interface Dictionary {
   
   public static final String INTERNAL_SEPARATOR = "\u0000";
   
-  public Keyword getWord(int wordId);
+  public Keyword getWord(long wordId);
   
   public List<Term> lookup(char[] chars, int off, int len) throws IOException;
 }
